Fix invalid data type for latlong

+++ src/main/kotlin/model/request/ClientRequest.kt
@@ -7,8 +7,8 @@ data class CreateClientRequest(
     val nfc: String,
     val name: String,
     val address: String? = null,
-    val latitude: Double? = null,
-    val longitude: Double? = null,
+    val latitude: Float? = null,
+    val longitude: Float? = null,
 )
 
 @Serializable
@@ -16,13 +16,13 @@ data class UpdateClientRequest(
     val nfc: String? = null,
     val name: String? = null,
     val address: String? = null,
-    val latitude: Double? = null,
-    val longitude: Double? = null,
+    val latitude: Float? = null,
+    val longitude: Float? = null,
 )
 
 @Serializable
 data class SetAddressRequest(
     val address: String,
-    val latitude: Double? = null,
-    val longitude: Double? = null,
+    val latitude: Float? = null,
+    val longitude: Float? = null,
 )
